refactor: OAuth 1.0a コアと Zaim 固有実装を分離し、test.extend フィクスチャに移行
- oauth1.ts を汎用 OAuth 1.0a コア (buildOAuth1AuthorizationHeader) のみに絞る
- zaim-oauth.ts を新設し Zaim エンドポイント固有のロジックを集約
- test-fixtures.ts を新設し oauthTest (test.extend) と RFC テストベクターを共有化
- oauth1.test.ts・zaim-oauth.test.ts を test.extend フィクスチャパターンに書き直し
