+++ apps/server/src/oauth1.ts
@@ -1,32 +1,14 @@
 /**
- * OAuth 1.0a implementation using Web Crypto API (Cloudflare Workers 対応)
+ * OAuth 1.0a コア実装 (Web Crypto API 使用、Cloudflare Workers 対応)
  *
- * Zaim API の OAuth フロー:
- *   1. Request Token 取得: POST https://api.zaim.net/v2/auth/request
- *   2. ユーザー認可: https://auth.zaim.net/users/auth?oauth_token=...
- *   3. Access Token 取得: POST https://api.zaim.net/v2/auth/access
+ * 特定サービスに依存しない汎用 OAuth 1.0a Authorization ヘッダー生成。
  */
-
-const ZAIM_REQUEST_TOKEN_URL = "https://api.zaim.net/v2/auth/request";
-const ZAIM_AUTHORIZE_URL = "https://auth.zaim.net/users/auth";
-const ZAIM_ACCESS_TOKEN_URL = "https://api.zaim.net/v2/auth/access";
 
 export interface OAuth1Config {
   consumerKey: string;
   consumerSecret: string;
   token?: string;
   tokenSecret?: string;
-}
-
-export interface RequestTokenResult {
-  oauthToken: string;
-  oauthTokenSecret: string;
-}
-
-export interface AccessTokenResult {
-  oauthToken: string;
-  oauthTokenSecret: string;
-  userId: string;
 }
 
 /** HMAC-SHA1 署名を Base64 エンコードして返す */
@@ -54,13 +36,13 @@ function pct(str: string): string {
 /**
  * OAuth 1.0a の Authorization ヘッダーを生成する
  *
- * @param method - HTTP メソッド (GET / POST)
- * @param baseUrl - クエリ文字列を含まないリクエスト URL
- * @param config - Consumer Key/Secret とオプションのトークン
- * @param extraOAuthParams - oauth_callback / oauth_verifier など追加 OAuth パラメータ
+ * @param method            - HTTP メソッド (GET / POST)
+ * @param baseUrl           - クエリ文字列を含まないリクエスト URL
+ * @param config            - Consumer Key/Secret とオプションのアクセストークン
+ * @param extraOAuthParams  - oauth_callback / oauth_verifier など追加 OAuth パラメータ
  * @param bodyOrQueryParams - 署名対象に含める本文またはクエリパラメータ
  */
-async function buildAuthorizationHeader(
+export async function buildOAuth1AuthorizationHeader(
   method: string,
   baseUrl: string,
   config: OAuth1Config,
@@ -80,7 +62,7 @@ async function buildAuthorizationHeader(
     oauthParams.oauth_token = config.token;
   }
 
-  // 署名ベース文字列用にすべてのパラメータをソートして結合
+  // 署名ベース文字列: すべてのパラメータをソートして結合
   const allParams = { ...oauthParams, ...bodyOrQueryParams };
   const paramString = Object.entries(allParams)
     .sort(([a], [b]) => (a < b ? -1 : a > b ? 1 : 0))
@@ -92,102 +74,10 @@ async function buildAuthorizationHeader(
 
   oauthParams.oauth_signature = await hmacSha1Base64(signingKey, baseString);
 
-  // Authorization ヘッダー値を生成
   const headerValue = Object.entries(oauthParams)
     .sort(([a], [b]) => (a < b ? -1 : a > b ? 1 : 0))
     .map(([k, v]) => `${pct(k)}="${pct(v)}"`)
     .join(", ");
 
   return `OAuth ${headerValue}`;
-}
-
-/**
- * Zaim から Request Token を取得する
- * @param config - Consumer Key/Secret
- * @param callbackUrl - OAuth 認可後のコールバック URL
- */
-export async function fetchZaimRequestToken(
-  config: OAuth1Config,
-  callbackUrl: string,
-): Promise<RequestTokenResult> {
-  const authHeader = await buildAuthorizationHeader("POST", ZAIM_REQUEST_TOKEN_URL, config, {
-    oauth_callback: callbackUrl,
-  });
-
-  const response = await fetch(ZAIM_REQUEST_TOKEN_URL, {
-    method: "POST",
-    headers: { Authorization: authHeader },
-  });
-
-  if (!response.ok) {
-    const body = await response.text();
-    throw new Error(`Request token failed [${response.status}]: ${body}`);
-  }
-
-  const params = new URLSearchParams(await response.text());
-  const oauthToken = params.get("oauth_token");
-  const oauthTokenSecret = params.get("oauth_token_secret");
-
-  if (!oauthToken || !oauthTokenSecret) {
-    throw new Error("Invalid request token response: missing oauth_token or oauth_token_secret");
-  }
-
-  return { oauthToken, oauthTokenSecret };
-}
-
-/** Zaim ユーザー認可 URL を構築する */
-export function buildZaimAuthorizeUrl(oauthToken: string): string {
-  return `${ZAIM_AUTHORIZE_URL}?oauth_token=${encodeURIComponent(oauthToken)}`;
-}
-
-/**
- * Request Token と Verifier を使って Access Token を取得する
- * @param config - Consumer Key/Secret + Request Token / Secret
- * @param oauthVerifier - Zaim から受け取った oauth_verifier
- */
-export async function fetchZaimAccessToken(
-  config: OAuth1Config,
-  oauthVerifier: string,
-): Promise<AccessTokenResult> {
-  const authHeader = await buildAuthorizationHeader("POST", ZAIM_ACCESS_TOKEN_URL, config, {
-    oauth_verifier: oauthVerifier,
-  });
-
-  const response = await fetch(ZAIM_ACCESS_TOKEN_URL, {
-    method: "POST",
-    headers: { Authorization: authHeader },
-  });
-
-  if (!response.ok) {
-    const body = await response.text();
-    throw new Error(`Access token failed [${response.status}]: ${body}`);
-  }
-
-  const params = new URLSearchParams(await response.text());
-  const oauthToken = params.get("oauth_token");
-  const oauthTokenSecret = params.get("oauth_token_secret");
-  const userId = params.get("user_id");
-
-  if (!oauthToken || !oauthTokenSecret || !userId) {
-    throw new Error("Invalid access token response: missing required fields");
-  }
-
-  return { oauthToken, oauthTokenSecret, userId };
-}
-
-/**
- * 保存済みアクセストークンを使って Zaim API リクエスト用の Authorization ヘッダーを生成する
- *
- * @param config - Consumer Key/Secret + Access Token / Secret
- * @param method - HTTP メソッド
- * @param url - クエリ文字列を含まないリクエスト URL
- * @param queryParams - 署名対象のクエリパラメータ
- */
-export async function buildZaimApiAuthHeader(
-  config: OAuth1Config,
-  method: string,
-  url: string,
-  queryParams: Record<string, string> = {},
-): Promise<string> {
-  return buildAuthorizationHeader(method, url, config, {}, queryParams);
 }
